Makefile Ok

+++ Sources/StrategyIpAddress.cpp
@@ -24,7 +24,7 @@ int             Plazza::StrategyIpAddress::FindInformation(int pos)
     if (!this->IsValidInformation(this->_fileContent[pos]))
         pos++;
     save = pos;
-    if (pos + 1 < this->_fileContent.length() && this->_fileContent[pos] == '0' && this->IsValidInformation(this->_fileContent[pos + 1]))
+    if (static_cast<unsigned int>(pos + 1) < this->_fileContent.length() && this->_fileContent[pos] == '0' && this->IsValidInformation(this->_fileContent[pos + 1]))
         val--;
     if (pos > 0 && !this->IsValidInformation(this->_fileContent[pos - 1]) && this->IsValidInformation(this->_fileContent[pos]) && atoi(&this->_fileContent[pos]) <= 255)
         val++;
@@ -34,7 +34,7 @@ int             Plazza::StrategyIpAddress::FindInformation(int pos)
     {
         if (this->_fileContent[pos] == '.')
             pos++;
-        if (pos + 1 < this->_fileContent.length() && this->_fileContent[pos] == '0' && this->IsValidInformation(this->_fileContent[pos + 1]))
+        if (static_cast<unsigned int>(pos + 1) < this->_fileContent.length() && this->_fileContent[pos] == '0' && this->IsValidInformation(this->_fileContent[pos + 1]))
             val--;
         if (pos > 0 && this->_fileContent[pos - 1] == '.' && this->IsValidInformation(this->_fileContent[pos]) && atoi(&this->_fileContent[pos]) <= 255)
             val++;
